# api/src/main/resources/assets/openapi.yml
...
openapi: 3.0.2
info:
  title: Marquez
  version: 0.51.2-SNAPSHOT
  description: Marquez is an open source **metadata service** for the **collection**, **aggregation**,
    and **visualization** of a data ecosystem's metadata.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
    description: Current API server
paths:
...
  /api/v1/sources/{source}:
    parameters:
      - $ref: '#/components/parameters/source'
    put:
      operationId: putSource
      summary: Create a source
      description: Creates a new source object. A source is the physical location of a dataset such as
        a table in PostgreSQL, or topic in Kafka. A source enables the grouping of physical datasets
        to their physical source.
      deprecated: true
      tags:
        - Sources
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatedSource'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
    get:
      operationId: getSource
      summary: Retrieve a source
      description: Returns a source.
      tags:
        - Sources
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Source'
...
  /api/v1/namespaces/{namespace}/datasets/{dataset}:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/dataset'
    put:
      operationId: putDataset
      summary: Create a dataset
      description: Creates a new dataset.
      deprecated: true
      tags:
        - Datasets
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
                - $ref: '#/components/schemas/DbTable'
                - $ref: '#/components/schemas/Stream'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
    get:
      operationId: getDataset
      summary: Retrieve a dataset
      description: Returns a dataset.
      tags:
        - Datasets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
    delete:
      operationId: deleteDataset
      summary: Soft deletes dataset.
      description: Soft deletes dataset. It will be un-deleted if new OpenLineage event containing this dataset comes.
      tags:
        - Datasets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
...
  /api/v1/namespaces/{namespace}/datasets:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/dataset'
    get:
      operationId: getDatasets
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all datasets
      description: Returns a list of datasets.
      tags:
        - Datasets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetList'
  /api/v1/namespaces/{namespace}/datasets/{dataset}/tags/{tag}:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/dataset'
      - $ref: '#/components/parameters/tag'
    post:
      operationId: addTagToDataset
      summary: Tag a dataset
      description: Tag an existing dataset.
      tags:
        - Datasets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
  /api/v1/namespaces/{namespace}/datasets/{dataset}/fields/{field}/tags/{tag}:
    parameters:
      - $ref: '#/components/parameters/namespace'
      - $ref: '#/components/parameters/dataset'
      - $ref: '#/components/parameters/field'
      - $ref: '#/components/parameters/tag'
    post:
      operationId: addTagToFieldOfDataset
      summary: Tag a field
      description: Tag an existing field of a dataset.
      tags:
        - Datasets
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
...
  /api/v1/lineage:
    post:
      operationId: recordLineage
      summary: Record a single lineage event
      description: Receive, process, and store lineage metadata using the [OpenLineage](https://github.com/OpenLineage/OpenLineage/blob/main/spec/OpenLineage.json) standard.
      tags:
        - Lineage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LineageEvent'
      responses:
        '200':
          description: OK
    get:
      operationId: getLineage
      parameters:
        - $ref: '#/components/parameters/nodeId'
        - $ref: '#/components/parameters/depth'
        - $ref: '#/components/parameters/aggregateToParentRun'
        - $ref: '#/components/parameters/includeFacets'
      tags:
        - Lineage
      summary: Get a lineage graph
      description: Returns a lineage graph for the specified node. Optionally aggregate child runs to parent and filter facets to prevent large response sizes.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageGraph'
...
  /api/v1/column-lineage:
    get:
      operationId: getColumnLineage
      parameters:
        - $ref: '#/components/parameters/nodeId'
        - $ref: '#/components/parameters/depth'
        - $ref: '#/components/parameters/withDownstream'
      tags:
        - Column lineage
      summary: Get a column lineage graph
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LineageGraph'
...
  /api/v2/namespaces/{namespace}/datasets:
    get:
      operationId: listDatasetsV2
      parameters:
        - $ref: '#/components/parameters/namespace'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/includeFacets'
      summary: List all datasets (V2)
      description: Returns a list of datasets for a namespace using service models.
      tags:
        - Datasets V2
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dataset'
  /api/v2/namespaces/{namespace}/datasets/{dataset}:
    get:
      operationId: getDatasetV2
      parameters:
        - $ref: '#/components/parameters/namespace'
        - $ref: '#/components/parameters/dataset'
        - $ref: '#/components/parameters/includeFacets'
      summary: Retrieve a dataset (V2)
      description: Returns a dataset using service models.
      tags:
        - Datasets V2
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
...
  /api/v3/namespaces/{namespace}/datasets:
    get:
      operationId: getNamespaceDatasetsV3
      parameters:
        - $ref: '#/components/parameters/namespace'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      summary: List all datasets for a namespace (V3 Graph)
      description: Returns a list of datasets for a namespace using the V3 Graph backend.
      tags:
        - Datasets V3
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasetList'
...
components:
  parameters:
...
    namespace:
      name: namespace
      in: path
      description: The name of the namespace.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-namespace
    source:
      name: source
      in: path
      description: The name of the source.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-source
    dataset:
      name: dataset
      in: path
      description: The name of the dataset.
      required: true
      schema:
        type: string
        maxLength: 1024
        example: my-dataset
    tag:
      name: tag
      in: path
      description: The name of the tag.
      required: true
      schema:
        type: string
        example: SENSITIVE
    field:
      name: field
      in: path
      description: The name of the field.
      required: true
      schema:
        type: string
        example: my_field
...
    limit:
      name: limit
      in: query
      description: The number of results to return from offset.
      required: false
      schema:
        type: integer
        example: 25
        default: 100
    offset:
      name: offset
      in: query
      description: The initial position from which to return results.
      required: false
      schema:
        type: integer
        default: 0
    nodeId:
      name: nodeId
      in: query
      schema:
        type: string
        example: dataset:food_delivery:public.delivery_7_days
      description: The ID of the node. A node can either be a dataset node, a dataset field node or a job node.
        The format of nodeId for dataset is `dataset:<namespace_of_dataset>:<name_of_the_dataset>`, for dataset field
        is `datasetField:<namespace_of_dataset>:<name_of_the_dataset>:<name_of_field>`, and for job is
        `job:<namespace_of_the_job>:<name_of_the_job>`.
      required: true
    depth:
      name: depth
      in: query
      schema:
        type: integer
        default: 20
      description: Depth of lineage graph to create.
      required: false
    withDownstream:
      name: withDownstream
      in: query
      schema:
        type: boolean
        default: false
      description: Determines if downstream lineage should be returned.
      required: false
    aggregateToParentRun:
      name: aggregateToParentRun
      in: query
      schema:
        type: boolean
        default: false
      description: When true, aggregates all child runs to their parent run UUID, grouping facets and lineage data together. Useful for jobs with dynamic child tasks.
      required: false
    includeFacets:
      name: includeFacets
      in: query
      schema:
        type: array
        items:
          type: string
        example: ["spark", "processing_engine"]
      description: Filters run facets to include only the specified facet names. Reduces response size and prevents PostgreSQL 256MB JSONB limit issues with large lineage queries. Facets are filtered at SQL level for optimal performance.
      required: false
      style: form
      explode: true
...
  schemas:
...
    CreatedSource:
      type: object
      properties:
        type:
          description: The type of the source.
          type: string
        connectionUrl:
          description: The URL to the location of the source.
          type: string
          format: URL
        description:
          description:  The description of the source.
          type: string
      required:
        - type
        - connectionUrl
      example:
        type: POSTGRESQL
        connectionUrl: jdbc:postgresql://db.example.com/mydb
        description: My first source!
    Source:
      type: object
      properties:
        type:
          description: The type of the source.
          type: string
        name:
          description: The name of the source.
          type: string
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the source was created.
          type: string
          format: date-time
        updatedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the source was updated.
          type: string
          format: date-time
        connectionUrl:
          description: The URL to the location of the source.
          type: string
          format: URL
        description:
          description:  The description of the dataset.
          type: string
      example:
        type: POSTGRESQL
        name: my-source
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        connectionUrl: jdbc:postgresql://db.example.com/mydb
        description: My first source!
...
    DbTable:
      type: object
      properties:
        type:
          description: The type of the dataset.
          type: string
          enum: [DB_TABLE]
        physicalName:
          description: The **physical** name of the table.
          type: string
        sourceName:
          description: The name of the source associated with the table.
          type: string
        fields:
          description: The fields of the table.
          type: array
          items:
            type: object
            properties:
              name:
                description: The name of the field.
                type: string
              type:
                description: The data type of the field.
                type: string
              tags:
                description: List of tags.
                type: string
              description:
                description: The description of the field.
                type: string
            required:
              - name
              - type
        tags:
          description: List of tags.
          type: array
          items:
            type: string
        description:
          description:  The description of the table.
          type: string
        runId:
          description: The ID associated with the run modifying the table.
          type: string
      required:
        - type
        - physicalName
        - sourceName
        - fields
      example:
        type: DB_TABLE
        physicalName: public.mytable
        sourceName: my-source
        fields: [
          {name: a, type: INTEGER},
          {name: b, type: TIMESTAMP},
          {name: c, type: INTEGER},
          {name: d, type: INTEGER}
        ]
        description: "My first dataset!"
    Stream:
      type: object
      properties:
        type:
          description: The type of the dataset.
          type: string
          enum: [STREAM]
        physicalName:
          description: The **physical** name of the stream.
          type: string
        sourceName:
          description: The name of the source associated with the stream.
          type: string
        schemaLocation:
          description: The URL of the schema for the stream data.
          type: string
        fields:
          description: The schema fields for the stream data.
          type: array
          items:
            type: object
            properties:
              name:
                description: The name of the field.
                type: string
              type:
                description: The type of the field.
                type: string
              tags:
                description: List of tags.
                type: array
                items:
                  type: string
              description:
                description: The description of the field.
                type: string
            required:
              - name
              - type
        tags:
          description: List of tags.
          type: array
          items:
            type: string
        description:
          description: The description of the stream.
          type: string
        runId:
          description: The ID associated with the run modifying the stream.
          type: string
      required:
        - type
        - physicalName
        - sourceName
        - fields
        - schemaLocation
      example:
        type: STREAM
        physicalName: my-stream
        sourceName: my-stream-source
        schemaLocation: 'http://registry.example.com/schemas/ids/1'
        fields: []
        description: "My first stream dataset!"
    DatasetId:
      type: object
      description: The ID of the dataset.
      properties:
        namespace:
          type: string
          description: The namespace of the dataset.
        name:
          type: string
          description: The name of the dataset.
    Dataset:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/DatasetId'
        type:
          description: The type of the dataset.
          type: string
        name:
          description: The **logical** name of the dataset.
          type: string
        physicalName:
          description: The **physical** name of the dataset.
          type: string
        createdAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the dataset was created.
          type: string
          format: date-time
        updatedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the dataset was updated.
          type: string
          format: date-time
        namespace:
          type: string
          description: The namespace of the dataset.
        sourceName:
          description: The name of the source associated with the dataset.
          type: string
        fields:
          description: The fields of the dataset.
          type: array
          items:
            type: object
            properties:
              name:
                description: The name of the field.
                type: string
              type:
                description: The data type of the field.
                type: string
              tags:
                description: List of tags.
                type: array
                items:
                  type: string
              description:
                description: The description of the field.
                type: string
            required:
              - name
              - type
        tags:
          description: List of tags.
          type: array
          items:
            type: string
        lastModifiedAt:
          description: An [ISO-8601](https://en.wikipedia.org/wiki/ISO_8601) timestamp representing the date/time the dataset was last modified by a successful run.
          type: string
          format: date-time
        lastLifecycleState:
          description: The last lifecycle state of the dataset.
          type: string
        description:
          description: The description of the dataset.
          type: string
        facets:
          $ref: '#/components/schemas/DatasetFacets'
        currentVersion:
          description: The current version of the dataset.
          type: string
          format: uuid
        deleted:
          description: The deleted state of the dataset.
          type: boolean
      example:
        id:
          namespace: my-namespace
          name: my-dataset
        type: DB_TABLE
        name: my-dataset
        physicalName: public.mytable
        createdAt: 2019-05-09T19:49:24.201361Z
        updatedAt: 2019-05-09T19:49:24.201361Z
        namespace: my-namespace
        sourceName: my-source
        fields: [
          {name': a, type: INTEGER, tags: []},
          {name': b, type: TIMESTAMP, tags: []},
          {name': c, type: INTEGER, tags: []},
          {name': d, type: INTEGER, tags: []}
        ]
        tags: []
        lastModifiedAt: null
        description: "My first dataset!"
        facets: {}
        currentVersion: "b1d626a2-6d3a-475e-9ecf-943176d4a8c6"
    DatasetList:
      type: object
      properties:
        datasets:
          items:
            $ref: '#/components/schemas/Dataset'
        totalCount:
          type: number
          description: The total number of datasets for the given namespace
      required:
        - datasets
...
    DatasetFacets:
      description: The facets of the dataset. For a full list of standard dataset facets, see [OpenLineage](https://github.com/OpenLineage/OpenLineage/blob/main/spec/OpenLineage.md#standard-facets).
      type: object
      additionalProperties:
        $ref: '#/components/schemas/CustomFacet'
...
    CustomFacet:
      description: A custom facet enables the extension of _dataset_, _job_, and _run_ metadata. A custom facet **must** also have a schema,
        where a version of the schema is identifiable via a URL. A field within the schema **must** not start with an underscore (`_`).
      allOf:
        - $ref: '#/components/schemas/BaseFacet'
        - type: object
          additionalProperties: true
    BaseFacet:
      description: All base facet fields are prefixed with an underscore (`_`) to avoid field naming conflicts defined with other facets.
      type: object
      properties:
        _producer:
          description: URI identifying the producer of this metadata. For example this could be a git url with a given tag or sha
          type: string
          format: uri
          example: https://github.com/OpenLineage/OpenLineage/blob/v1-0-0/client
        _schemaURL:
          description: The URL to the corresponding version of the schema definition following a $ref URL Reference (see https://swagger.io/docs/specification/using-ref/)
          type: string
          format: uri
          example: https://github.com/OpenLineage/OpenLineage/blob/v1-0-0/spec/OpenLineage.yml#MyCustomJobFacet
      required:
        - _producer
        - _schemaURL
    LineageGraph:
      type: object
      properties:
        graph:
          type: array
          items:
            $ref: '#/components/schemas/GraphNode'
    GraphNode:
      type: object
      properties:
        id:
          type: string
          description: The ID of the node.
        type:
          $ref: '#/components/schemas/NodeType'
        data:
          type: object
          description: The data of the node.
          discriminator:
            propertyName: type
            mapping:
              'DATASET': '#/components/schemas/Dataset'
              'JOB': '#/components/schemas/Job'
          oneOf:
            - $ref: '#/components/schemas/Dataset'
            - $ref: '#/components/schemas/Job'
        inEdges:
          type: array
          description: The in edges of the node.
          items:
            $ref: '#/components/schemas/GraphEdge'
        outEdges:
          type: array
          description: The out edges of the node.
          items:
            $ref: '#/components/schemas/GraphEdge'
    NodeType:
      type: string
      enum: [ JOB, DATASET ]
      description: The type of the diff node.
    GraphEdge:
      type: object
      properties:
        origin:
          description: The ID of the origin node.
          type: string
        destination:
          description: The ID of the destination node.
          type: string
...
    LineageEvent:
      example:
        eventType: COMPLETE
        eventTime: 2019-05-09T19:49:24.201361Z
        run: {runId: d46e465b-d358-4d32-83d4-df660ff614dd}
        job: {namespace: my-namespace,name: my-job}
        outputs: [
          {
            namespace: my-namespace,
            name: my-output,
            facets: {
              schema: {
                _producer: "https://github.com/OpenLineage/OpenLineage/blob/v1-0-0/client",
                _schemaURL: "https://github.com/OpenLineage/OpenLineage/blob/v1-0-0/spec/OpenLineage.json#/definitions/SchemaDatasetFacet",
                fields: [
                  { name: a, type: INTEGER },
                  { name: b, type: TIMESTAMP },
                  { name: c, type: INTEGER },
                  { name: d, type: INTEGER }
                ]
              }
            }
          }
        ]
        producer: "https://github.com/OpenLineage/OpenLineage/blob/v1-0-0/client"
...